refactor - put everything in its own auth package
