+++ refiner/app/main.py
@@ -3,38 +3,26 @@ import os
 from contextlib import asynccontextmanager
 from pathlib import Path
 
-from authlib.integrations.starlette_client import OAuth
-from fastapi import APIRouter, FastAPI, Request
+from fastapi import APIRouter, FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse, RedirectResponse
 from fastapi.staticfiles import StaticFiles
 from starlette.middleware.sessions import SessionMiddleware
 
+from .api.auth.config import SESSION_SECRET_KEY
+from .api.auth.handlers import auth_router
 from .api.middleware.spa import SPAFallbackMiddleware
 from .api.v1.demo import run_expired_file_cleanup_task
 from .api.v1.v1_router import router as v1_router
 from .core.app.base import BaseService
 from .core.app.openapi import create_custom_openapi
-from .core.config import ENVIRONMENT
-
-SECRET_KEY = "super-secret-key"
 
 # environment configuration
 is_production = os.getenv("PRODUCTION", "false").lower() == "true"
 
 # create router
 router = APIRouter(prefix="/api")
+router.include_router(auth_router)
 router.include_router(v1_router)
-
-
-oauth = OAuth()
-oauth.register(
-    name=ENVIRONMENT["AUTH_PROVIDER"],
-    client_id=ENVIRONMENT["AUTH_CLIENT_ID"],
-    client_secret=ENVIRONMENT["AUTH_CLIENT_SECRET"],
-    server_metadata_url=f"{ENVIRONMENT['AUTH_ISSUER']}/.well-known/openid-configuration",
-    client_kwargs={"scope": "openid email profile"},
-)
 
 
 # define health check endpoint at the service level
@@ -49,104 +37,6 @@ async def health_check() -> dict[str, str]:
     """
 
     return {"status": "OK"}
-
-
-@router.get("/login")
-async def login(request: Request) -> RedirectResponse:
-    """
-    Initiates the OAuth2 login flow by redirecting the user to the authorization endpoint.
-
-    Args:
-        request (Request): The incoming HTTP request.
-
-    Returns:
-        RedirectResponse: A redirect response that sends the user to the OAuth provider's login page.
-    """
-    redirect_uri = "http://localhost:8080/api/auth/callback"
-    return await oauth.keycloak.authorize_redirect(request, redirect_uri)
-
-
-@router.get("/auth/callback")
-async def auth_callback(request: Request) -> dict[str, str]:
-    """
-    Handles the OAuth2 callback by exchanging the authorization code for tokens, parsing the ID token, and returning the user information.
-
-    Args:
-        request (Request): The incoming HTTP request containing the authorization code.
-
-    Returns:
-        dict[str, str]: A dictionary of user claims extracted from the ID token.
-
-    Raises:
-        Exception: If token exchange or ID token parsing fails.
-    """
-
-    try:
-        token = await oauth.keycloak.authorize_access_token(request)
-        nonce = request.session.get("nonce")
-        user = await oauth.keycloak.parse_id_token(token, nonce)
-
-        request.session["id_token"] = token["id_token"]
-        request.session["user"] = user
-
-        # print(dict(user))
-
-        return RedirectResponse(url="http://localhost:8081")
-    except Exception as e:
-        print("Callback error:", e)
-        raise e
-
-
-@router.get("/user")
-async def get_user(request: Request) -> JSONResponse:
-    """
-    Returns the current logged-in user's information.
-
-    Reads user info from the session or token.
-
-    Returns:
-        JSON object with user claims if authenticated.
-
-    Raises:
-        HTTPException 401 if user not authenticated.
-    """
-    user = request.session.get("user")
-    if not user:
-        return JSONResponse(content=user)
-    return JSONResponse(content=user)
-
-
-@router.get("/logout")
-async def logout(request: Request) -> RedirectResponse:
-    """
-    Logs the user out by clearing the session and redirecting to the auth provider logout endpoint.
-
-    Args:
-        request (Request): The incoming HTTP request.
-
-    Returns:
-        RedirectResponse: A redirect to the auth provider logout endpoint and back to the frontend.
-    """
-
-    # Redirect to client
-    post_logout_redirect_uri = "http://localhost:8081"
-
-    id_token = request.session.get("id_token")
-    if not id_token:
-        # Fallback if user is not logged in
-        return RedirectResponse(post_logout_redirect_uri)
-
-    # Clear the session
-    request.session.clear()
-
-    # Logout from auth provider
-    auth_provider_logout_url = (
-        "http://localhost:8082/realms/refiner/protocol/openid-connect/logout"
-        f"?post_logout_redirect_uri=http://localhost:8081"
-        f"&id_token_hint={id_token}"
-    )
-
-    return RedirectResponse(url=auth_provider_logout_url)
 
 
 @asynccontextmanager
@@ -187,4 +77,4 @@ app.add_middleware(
     allow_headers=["*"],
 )
 app.add_middleware(SPAFallbackMiddleware)
-app.add_middleware(SessionMiddleware, secret_key=SECRET_KEY)
+app.add_middleware(SessionMiddleware, secret_key=SESSION_SECRET_KEY)
